add load command
TKM uses this script to create and delete clusters. But it was doing a
`kind load docker-image' locally, which requires DOCKER_HOST and
KIND_EXPERIMENTAL_PROVIDER to be managed outside this script just to
load. So extend this script to load images into cluster so podman
management is mostly isolated.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>